Write response body when topic creation fails
When topic creation fails with 400 Bad Request, no information is available as to why.  ASB returns a reason message in the body as JSON.  This commit causes that reason to be printed out in the shell window when the user hits that issue (e.g. if namespace partioning is not set up correctly in ASB).